+++ CHANGELOG.md
@@ -5,11 +5,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `Field.Checkbox`.
-- Feature: Add `customSize` props to `HorizontalStroke`.
-- Feature: Add `HorizontalStroke` size CSS: `TINY`, `DEFAULT`, `BIG` and `HUGE`.
 - Feature: Export `FLUID`, `DEFAULT`, `TINY` and `BIG` css from `Button`.
-- Feature: Add `IconList`.
 - Fix: Update selected option style to `Select` component.
+
+## [1.22.0] - 2019-03-18
+
+Features:
+- Add `customSize` props to `HorizontalStroke`.
+- Add `HorizontalStroke` size CSS: `TINY`, `DEFAULT`, `BIG` and `HUGE`.
+- Add `IconList`.
+- Add empty `Prev` and `Next` container to `NavTabCarousel` when
+  `NavTabCarousel.Prev` or `NavTabCarousel.Next` are empty.
 
 ## [1.21.0] - 2019-03-11